refactor(table): extract reusable dashboard table components and make manage-users prop-driven
