+++ src/features/manage-users/components/index.ts
@@ -1,3 +1,3 @@
 export * from "./location-search-dropdown";
-export * from "./ManagementTablePage";
+export * from "../../../components/dashboard/table/ManagementTablePage";
 export * from "./multi-select-dropdown";
